[Decomp][World] Cohesion split: extract config + session management to World*.cpp
