+++ src/game/WorldHandlers/WorldSessionMgr.cpp
@@ -1,0 +1,321 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+#include "World.h"
+#include "WorldSession.h"
+#include "WorldPacket.h"
+#include "Opcodes.h"
+#include "Player.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+
+/**
+ * @file WorldSessionMgr.cpp
+ * @brief Cohesion split of World.cpp -- session management: zone player lookup, session add/remove and the login wait-queue. Same World class; no behaviour change. CMake file(GLOB) picks this file up automatically; World.h is unchanged.
+ */
+
+/// Find a player in a specified zone
+Player* World::FindPlayerInZone(uint32 zone)
+{
+    ///- circle through active sessions and return the first player found in the zone
+    SessionMap::const_iterator itr;
+    for (itr = m_sessions.begin(); itr != m_sessions.end(); ++itr)
+    {
+        if (!itr->second)
+        {
+            continue;
+        }
+        Player* player = itr->second->GetPlayer();
+        if (!player)
+        {
+            continue;
+        }
+        if (player->IsInWorld() && player->GetZoneId() == zone)
+        {
+            // Used by the weather system. We return the player to broadcast the change weather message to him and all players in the zone.
+            return player;
+        }
+    }
+    return NULL;
+}
+
+/// Find a session by its id
+WorldSession* World::FindSession(uint32 id) const
+{
+    SessionMap::const_iterator itr = m_sessions.find(id);
+
+    if (itr != m_sessions.end())
+    {
+        return itr->second;
+    }                                 // also can return NULL for kicked session
+    else
+    {
+        return NULL;
+    }
+}
+
+/// Remove a given session
+bool World::RemoveSession(uint32 id)
+{
+    ///- Find the session, kick the user, but we can't delete session at this moment to prevent iterator invalidation
+    SessionMap::const_iterator itr = m_sessions.find(id);
+
+    if (itr != m_sessions.end() && itr->second)
+    {
+        if (itr->second->PlayerLoading())
+        {
+            return false;
+        }
+        itr->second->KickPlayer();
+    }
+
+    return true;
+}
+
+/**
+ * @brief Queues a new session for addition during the world update.
+ *
+ * @param s The session to add.
+ */
+void World::AddSession(WorldSession* s)
+{
+    std::lock_guard<std::mutex> guard(m_sessionAddQueueLock);
+
+    m_sessionAddQueue.push_back(s);
+}
+
+void
+World::AddSession_(WorldSession* s)
+{
+    MANGOS_ASSERT(s);
+
+    // NOTE - Still there is race condition in WorldSession* being used in the Sockets
+
+    ///- kick already loaded player with same account (if any) and remove session
+    ///- if player is in loading and want to load again, return
+    if (!RemoveSession(s->GetAccountId()))
+    {
+        s->KickPlayer();
+        delete s;                                           // session not added yet in session list, so not listed in queue
+        return;
+    }
+
+    // decrease session counts only at not reconnection case
+    bool decrease_session = true;
+
+    // if session already exist, prepare to it deleting at next world update
+    // NOTE - KickPlayer() should be called on "old" in RemoveSession()
+    {
+        SessionMap::const_iterator old = m_sessions.find(s->GetAccountId());
+
+        if (old != m_sessions.end())
+        {
+            // prevent decrease sessions count if session queued
+            if (RemoveQueuedSession(old->second))
+            {
+                decrease_session = false;
+            }
+            // not remove replaced session form queue if listed
+            delete old->second;
+        }
+    }
+
+    m_sessions[s->GetAccountId()] = s;
+
+    uint32 Sessions = GetActiveAndQueuedSessionCount();
+    uint32 pLimit = GetPlayerAmountLimit();
+    uint32 QueueSize = GetQueuedSessionCount();             // number of players in the queue
+
+    // so we don't count the user trying to
+    // login as a session and queue the socket that we are using
+    if (decrease_session)
+    {
+        --Sessions;
+    }
+
+    if (pLimit > 0 && Sessions >= pLimit && s->GetSecurity() == SEC_PLAYER)
+    {
+        AddQueuedSession(s);
+        UpdateMaxSessionCounters();
+        DETAIL_LOG("PlayerQueue: Account id %u is in Queue Position (%u).", s->GetAccountId(), ++QueueSize);
+        return;
+    }
+
+    WorldPacket packet(SMSG_AUTH_RESPONSE, 17);
+
+    packet.WriteBit(false);                                 // has queue
+    packet.WriteBit(true);                                  // has account info
+
+    packet << uint32(0);                                    // Unknown - 4.3.2
+    packet << uint8(s->Expansion());                        // 0 - normal, 1 - TBC, 2 - WotLK, 3 - CT. must be set in database manually for each account
+    packet << uint32(0);                                    // BillingTimeRemaining
+    packet << uint8(s->Expansion());                        // 0 - normal, 1 - TBC, 2 - WotLK, 3 - CT. Must be set in database manually for each account.
+    packet << uint32(0);                                    // BillingTimeRested
+    packet << uint8(0);                                     // BillingPlanFlags
+    packet << uint8(AUTH_OK);
+
+    s->SendPacket(&packet);
+
+    s->SendAddonsInfo();
+
+    WorldPacket pkt(SMSG_CLIENTCACHE_VERSION, 4);
+    pkt << uint32(getConfig(CONFIG_UINT32_CLIENTCACHE_VERSION));
+    s->SendPacket(&pkt);
+
+    s->SendTutorialsData();
+
+    UpdateMaxSessionCounters();
+
+    // Updates the population
+    if (pLimit > 0)
+    {
+        float popu = float(GetActiveSessionCount());        // updated number of users on the server
+        popu /= pLimit;
+        popu *= 2;
+
+        static SqlStatementID id;
+
+        SqlStatement stmt = LoginDatabase.CreateStatement(id, "UPDATE `realmlist` SET `population` = ? WHERE `id` = ?");
+        stmt.PExecute(popu, realmID);
+
+        DETAIL_LOG("Server Population (%f).", popu);
+    }
+}
+
+/**
+ * @brief Returns the current queue position of a session.
+ *
+ * @param sess The session to locate.
+ * @return The 1-based queue position, or 0 if not queued.
+ */
+int32 World::GetQueuedSessionPos(WorldSession* sess)
+{
+    uint32 position = 1;
+
+    for (Queue::const_iterator iter = m_QueuedSessions.begin(); iter != m_QueuedSessions.end(); ++iter, ++position)
+        if ((*iter) == sess)
+        {
+            return position;
+        }
+
+    return 0;
+}
+
+/**
+ * @brief Places a session into the login queue and notifies the client.
+ *
+ * @param sess The session being queued.
+ */
+void World::AddQueuedSession(WorldSession* sess)
+{
+    sess->SetInQueue(true);
+    m_QueuedSessions.push_back(sess);
+
+    // The 1st SMSG_AUTH_RESPONSE needs to contain other info too.
+    WorldPacket packet (SMSG_AUTH_RESPONSE, 21);
+
+    packet.WriteBit(true);                                  // has queue
+    packet.WriteBit(false);                                 // unk queue-related
+    packet.WriteBit(true);                                  // has account data
+
+    packet << uint32(0);                                    // Unknown - 4.3.2
+    packet << uint8(sess->Expansion());                     // 0 - normal, 1 - TBC, 2 - WotLK, 3 - CT. must be set in database manually for each account
+    packet << uint32(0);                                    // BillingTimeRemaining
+    packet << uint8(sess->Expansion());                     // 0 - normal, 1 - TBC, 2 - WotLK, 3 - CT. Must be set in database manually for each account.
+    packet << uint32(0);                                    // BillingTimeRested
+    packet << uint8(0);                                     // BillingPlanFlags
+    packet << uint8(AUTH_WAIT_QUEUE);
+    packet << uint32(GetQueuedSessionPos(sess));            // position in queue
+
+    sess->SendPacket(&packet);
+}
+
+/**
+ * @brief Removes a session from the login queue and updates later positions.
+ *
+ * @param sess The session to remove.
+ * @return true if the session was queued and removed; otherwise false.
+ */
+bool World::RemoveQueuedSession(WorldSession* sess)
+{
+    // sessions count including queued to remove (if removed_session set)
+    uint32 sessions = GetActiveSessionCount();
+
+    uint32 position = 1;
+    Queue::iterator iter = m_QueuedSessions.begin();
+
+    // search to remove and count skipped positions
+    bool found = false;
+
+    for (; iter != m_QueuedSessions.end(); ++iter, ++position)
+    {
+        if (*iter == sess)
+        {
+            sess->SetInQueue(false);
+            iter = m_QueuedSessions.erase(iter);
+            found = true;                                   // removing queued session
+            break;
+        }
+    }
+
+    // iter point to next socked after removed or end()
+    // position store position of removed socket and then new position next socket after removed
+
+    // if session not queued then we need decrease sessions count
+    if (!found && sessions)
+    {
+        --sessions;
+    }
+
+    // accept first in queue
+    if ((!m_playerLimit || (int32)sessions < m_playerLimit) && !m_QueuedSessions.empty())
+    {
+        WorldSession* pop_sess = m_QueuedSessions.front();
+        pop_sess->SetInQueue(false);
+        pop_sess->SendAuthWaitQue(0);
+        pop_sess->SendAddonsInfo();
+
+        WorldPacket pkt(SMSG_CLIENTCACHE_VERSION, 4);
+        pkt << uint32(getConfig(CONFIG_UINT32_CLIENTCACHE_VERSION));
+        pop_sess->SendPacket(&pkt);
+
+        pop_sess->SendAccountDataTimes(GLOBAL_CACHE_MASK);
+        pop_sess->SendTutorialsData();
+
+        m_QueuedSessions.pop_front();
+
+        // update iter to point first queued socket or end() if queue is empty now
+        iter = m_QueuedSessions.begin();
+        position = 1;
+    }
+
+    // update position from iter to end()
+    // iter point to first not updated socket, position store new position
+    for (; iter != m_QueuedSessions.end(); ++iter, ++position)
+    {
+        (*iter)->SendAuthWaitQue(position);
+    }
+
+    return found;
+}
